+++ changelog.md
@@ -3,6 +3,7 @@ Changelog
 dev
 ------------------
 - Predicate.content attribute no longer includes an extra placeholder for a "quantity"
+- Predicate.content attribute must be a valid Python string template
 
 0.4.1 (2021-01-02)
 ------------------